# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLess Encrypted Cloud Sync API
  version: 1.0-draft.1
  description: 服务端 0.1.0 已实现并通过本地合同验证；尚未部署或完成真实 OAuth 联调。仅存储端到端加密快照。行为约束见 http-api.md，AAD 与密码学见 encryption-protocol.md。
servers:
- url: https://sync.openless.example
  description: 保留示例域名，不是生产地址
security:
- SyncBearer: []
paths:
...
  /v1/me/vault:
    get:
      operationId: getVaultMetadata
      summary: 读取当前账号元信息，空库也返回 200
      parameters:
      - name: If-None-Match
        in: header
        required: false
        schema:
          $ref: '#/components/schemas/StrongETag'
      responses:
        '200':
          description: 当前元信息及强 ETag
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            ETag:
              description: 必须返回，客户端按原样保存的强实体标签。
              schema:
                $ref: '#/components/schemas/StrongETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultMetadata'
        '304':
          description: 与该 ETag 一致，无响应体；保留本地已缓存的同账号元信息
          headers:
            ETag:
              schema:
                $ref: '#/components/schemas/StrongETag'
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Retry-After:
              description: 必须返回秒数；只阻止提前重试，不要求客户端启动轮询。
              schema:
                type: integer
                minimum: 1
                maximum: 86400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteVault
      summary: 按版本删除密文并保留墓碑，防止旧设备自动复活
      parameters:
      - name: If-Match
        in: header
        required: true
        description: 最近元信息的强 ETag，不允许通配符或弱标签。
        schema:
          $ref: '#/components/schemas/StrongETag'
      - name: Idempotency-Key
        in: header
        required: true
        description: 与 body.operationId 相同，重试必须复用原值与请求字节。
        schema:
          $ref: '#/components/schemas/UuidV4'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteVaultRequest'
      responses:
        '200':
          description: 删除回执；重复请求重放原回执
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Idempotency-Replayed:
              description: 必须返回；true 表示原操作的回执重放，不能假定它仍是最新头。
              schema:
                type: string
                enum:
                - 'true'
                - 'false'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationReceipt'
        '400':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '412':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '415':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '428':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Retry-After:
              description: 必须返回秒数；只阻止提前重试，不要求客户端启动轮询。
              schema:
                type: integer
                minimum: 1
                maximum: 86400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 结构化错误；按 code 处理，不能据 message 推断是否可覆盖数据。
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    GithubBearer:
      type: http
      scheme: bearer
      description: 仅 /v1/auth/github 使用；必须属于已配置的云同步 GitHub OAuth 应用。
    SyncBearer:
      type: http
      scheme: bearer
      description: 本服务签发、账号限定、最长 15 分钟的同步访问令牌。
  schemas:
    UuidV4:
      type: string
      format: uuid
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
    Revision:
      type: string
      pattern: ^(0|[1-9][0-9]{0,19})$
      description: 规范 uint64 十进制字符串；还须数值校验不超过 18446744073709551615。
    PositiveRevision:
      type: string
      pattern: ^[1-9][0-9]{0,19}$
      description: 正 uint64 十进制字符串；不可按浮点数处理。
    GithubId:
      type: string
      pattern: ^[1-9][0-9]{0,19}$
      description: 正 uint64 十进制字符串；不可按浮点数处理。
    Sha256:
      type: string
      pattern: ^[0-9a-f]{64}$
    StrongETag:
      type: string
      pattern: ^"[!#-~]{1,128}"$
      description: 包含双引号的强实体标签；客户端必须当成不透明字符串。
...
    VaultMetadata:
      type: object
      additionalProperties: false
      required:
      - protocolVersion
      - state
      - ownerGithubId
      - revision
      - vaultId
      - keyId
      - updatedAt
      - payloadSchemaVersion
      - ciphertextBytes
      - ciphertextSha256
      - lastOperationId
      properties:
        protocolVersion:
          type: integer
          const: 1
        state:
          type: string
          enum:
          - empty
          - active
          - deleted
        ownerGithubId:
          $ref: '#/components/schemas/GithubId'
        revision:
          $ref: '#/components/schemas/Revision'
        vaultId:
          oneOf:
          - $ref: '#/components/schemas/UuidV4'
          - type: 'null'
        keyId:
          oneOf:
          - $ref: '#/components/schemas/UuidV4'
          - type: 'null'
        updatedAt:
          oneOf:
          - type: string
            format: date-time
          - type: 'null'
        payloadSchemaVersion:
          oneOf:
          - type: integer
            const: 1
          - type: 'null'
        ciphertextBytes:
          type: integer
          minimum: 0
          maximum: 16777232
        ciphertextSha256:
          oneOf:
          - $ref: '#/components/schemas/Sha256'
          - type: 'null'
        lastOperationId:
          oneOf:
          - $ref: '#/components/schemas/UuidV4'
          - type: 'null'
      description: 空库也返回 200；state 条件定义见 oneOf。
      oneOf:
      - properties:
          state:
            const: empty
          revision:
            const: '0'
          vaultId:
            type: 'null'
          keyId:
            type: 'null'
          updatedAt:
            type: 'null'
          payloadSchemaVersion:
            type: 'null'
          ciphertextBytes:
            const: 0
          ciphertextSha256:
            type: 'null'
          lastOperationId:
            type: 'null'
      - properties:
          state:
            const: active
          revision:
            $ref: '#/components/schemas/PositiveRevision'
          vaultId:
            $ref: '#/components/schemas/UuidV4'
          keyId:
            $ref: '#/components/schemas/UuidV4'
          updatedAt:
            type: string
            format: date-time
          payloadSchemaVersion:
            const: 1
          ciphertextBytes:
            type: integer
            minimum: 65552
            maximum: 16777232
          ciphertextSha256:
            $ref: '#/components/schemas/Sha256'
          lastOperationId:
            $ref: '#/components/schemas/UuidV4'
      - properties:
          state:
            const: deleted
          revision:
            $ref: '#/components/schemas/PositiveRevision'
          vaultId:
            $ref: '#/components/schemas/UuidV4'
          keyId:
            type: 'null'
          updatedAt:
            type: string
            format: date-time
          payloadSchemaVersion:
            type: 'null'
          ciphertextBytes:
            const: 0
          ciphertextSha256:
            type: 'null'
          lastOperationId:
            $ref: '#/components/schemas/UuidV4'
...
    DeleteVaultRequest:
      type: object
      additionalProperties: false
      required:
      - protocolVersion
      - baseRevision
      - operationId
      - expectedVaultId
      properties:
        protocolVersion:
          type: integer
          const: 1
        baseRevision:
          $ref: '#/components/schemas/Revision'
        operationId:
          $ref: '#/components/schemas/UuidV4'
        expectedVaultId:
          $ref: '#/components/schemas/UuidV4'
    OperationReceipt:
      type: object
      additionalProperties: false
      required:
      - operationId
      - status
      - kind
      - committedRevision
      - committedAt
      - vaultId
      - ciphertextSha256
      properties:
        operationId:
          $ref: '#/components/schemas/UuidV4'
        status:
          type: string
          const: committed
        kind:
          type: string
          enum:
          - create
          - snapshot
          - password_change
          - delete
        committedRevision:
          $ref: '#/components/schemas/PositiveRevision'
        committedAt:
          type: string
          format: date-time
        vaultId:
          $ref: '#/components/schemas/UuidV4'
        ciphertextSha256:
          oneOf:
          - $ref: '#/components/schemas/Sha256'
          - type: 'null'
      oneOf:
      - properties:
          kind:
            const: delete
          ciphertextSha256:
            type: 'null'
      - properties:
          kind:
            enum:
            - create
            - snapshot
            - password_change
          ciphertextSha256:
            $ref: '#/components/schemas/Sha256'
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          - requestId
          properties:
            code:
              type: string
              enum:
              - invalid_request
              - unsupported_protocol
              - invalid_crypto_header
              - unauthenticated
              - session_expired
              - wrong_oauth_app
              - owner_mismatch
              - vault_empty
              - vault_deleted
              - operation_not_found
              - idempotency_key_reused
              - key_epoch_changed
              - vault_exists
              - revision_exhausted
              - revision_conflict
              - payload_too_large
              - unsupported_media_type
              - precondition_required
              - rate_limited
              - internal_error
              - service_unavailable
            message:
              type: string
              minLength: 1
              maxLength: 200
            requestId:
              $ref: '#/components/schemas/UuidV4'
            currentRevision:
              $ref: '#/components/schemas/Revision'
...